Moved application route reopening to mixin to avoid unreliable bahaviour
