+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ember-modals",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "directories": {
     "doc": "doc",
     "test": "tests"
@@ -17,7 +17,7 @@
   "engines": {
     "node": ">= 0.10.0"
   },
-  "author": "",
+  "author": "Duncan Walker <walkerdu@bc.edu>",
   "license": "MIT",
   "devDependencies": {
     "broccoli-asset-rev": "^2.0.0",
