+++ tests/dummy/app/routes/application.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import ApplicationRouteMixin from 'ember-modals/mixins/routes/application';
+
+export default Ember.Route.extend(
+  ApplicationRouteMixin, {
+
+});
